WIP: separate code/data generated by run-jsc-stress-tests
Instead of generating variations on the same script in test_script_NNN
and then executing them without any arguments, generate a simpler
testcase description that simply assigns values to a few variables,
namely:

DIRECTORY="..."
TESTNAME="..."
CMD="..."
OUTPUT_HANDLER="..."
ERROR_HANDLER="..."

(with special handling of exported environment variables, see the code
for details).

Then, provide a helper (dispatch-testcase) that executes the testcase
and invoke it like so:

sh ../.helpers/dispatch-testcase ./testcase_47 test $(FLAGS) 3f4e08ee962761f0925a2bfad06f8117

(where "test" is the action, FLAGS pass along --progress-meter,
--verbosity and --report-execution-time) and the hex value is the unique
run ID.

This has the following advantages:
  - We can easily diff the generated testcase files every time we change
    something in the run* function definitions in run-jsc-stress-tests
    to ensure the results match the intention.
  - There is no need for a reproduction command any more, because the
    dispatch-testcase script provides a "run" action to run the command
    interactively. We can also provide a "wrap" action to run the
    command under a debugger.
  - Importantly, it becomes trivial to implement one's own helpers cause
    it's easy enough to just source the testcase_NNN file and
    do your own thing. Great for one-off scripts when debugging
    something.

  - Instead of generating ruby code, we simply switch to
    dispatch-testcase.rb (or ask the playstation runner to execute the
    testcases). This gets rid of the significant code duplication in
    webkitruby/jsc-stress-test-writer-{default,ruby,playstation}.rb. It
    also means we can easily run the shellcheck/rubocop linters on the
    dispatchers and easily catch mistakes, whereas ATM it's awkward to do
    that for all combinations.

The only disadvantages I can think of:
  - Generating everything inline in each script means that the script is
    slightly faster. Testing with /bin/true as the command, executing
    the current script is 1.35x faster than the dispatcher when using the
    busybox sh (less on other shells that have an appreciable startup
    time). The difference for a full run of the jsc stress tests is in
    the noise in my setup.
  - Executing the script directly when in the .runner directory doesn't
    work out of the box anymore (though this could be implemented), so
    it may be a suprise when a developer first encounters this change.

The patch also introduces the jsc-stress-test-selftest directory. The
scripts there test all combinations of output/error handlers with all
combinations of test runners and writers, all kinds of test behaviors
that the handlers look for and for various levels of verbosity. The
directory contains the "golden" files with the output of the "default"
test writer, which everything adjusts to (previously, there were slight
inconsistencies in the behavior of the generated ruby code).

Given how important it is to not accidentally introduce behavioral
changes that invalidate JSC stress test results (in this or other
patches), I think extensive testing is called for.

A minor convenience is that the new dispatchers also don't need to be
run from .runner, for easier scripting.