Fix ARMv7 C ABI for Int64s https://bugs.webkit.org/show_bug.cgi?id=279974
Reviewed by NOBODY (OOPS!).

On ARMv7, Int64 arguments need to be passed on even register pairs (so,
r0/r1 or r2/r3, but not r1/r2). Pass in dummy Int32 arguments in
LowerInt64 so that AirCCallingConvention.cpp does not have to know about
the provenance of Int32 arguments and whether they were originally part
of an Int64.

* Source/JavaScriptCore/b3/B3CCallValue.h:
* Source/JavaScriptCore/b3/B3LowerInt64.cpp:
* Source/JavaScriptCore/b3/air/AirCCallingConvention.cpp:
(JSC::B3::Air::computeCCallingConvention):
(JSC::B3::Air::computeCCallArguments):